+++ css-inline/tests/test_inlining.rs
@@ -663,7 +663,7 @@ fn remote_network_stylesheet_invalid_url() {
 </html>"#;
     let error = inline(html).expect_err("Should fail");
     #[cfg(feature = "http")]
-    let expected = "Invalid base URL: http:";
+    let expected = "builder error: empty host: http:";
     #[cfg(not(feature = "http"))]
     let expected = "Loading external URLs requires the `http` feature";
     assert_eq!(error.to_string(), expected);